started adding tests, fixed circular deps
