+++ method/resources/__init__.py
@@ -1,11 +1,7 @@
 # type: ignore
-from method.resources.Accounts import Account, AccountResource, AccountPayoffsResource, \
-  AccountBalancesResource, AccountCardsResource, AccountSensitiveResource, AccountVerificationSessionResource, \
-  AccountSubscriptionsResource, AccountTransactionsResource, AccountUpdatesResource
+from method.resources.Accounts import Account, AccountResource
 from method.resources.Element import Element, ElementResource
-from method.resources.Entities import Entity, EntityResource, EntityConnectResource, \
-  EntityCreditScoresResource, EntityIdentityResource, EntityVerificationSessionResource, \
-  EntityProductResource, EntitySensitiveResource, EntitySubscriptionsResource
+from method.resources.Entities import Entity, EntityResource
 from method.resources.Merchant import Merchant, MerchantResource
 from method.resources.Payments import Payment, PaymentResource
 from method.resources.Report import Report, ReportResource
