+++ requirements.txt
@@ -1,1 +1,2 @@
 hammock==0.2.4
+pytest==8.2.1
